Create The strongest Wall 1
This is an optimization problem, in which dynamic programming can be applied. In this strategy, we indirectly check all the possibilities of the sum, which reduces time complexity for solving the problem.We are either including or excluding the element and making dp[i][j]=1, whenever that sum is possible either by including or excluding a particular element a[i-1].I actually solved by taking the minimum difference between 2 subsets of the sum. Then printed the minimum value of the 2 subsets.

Topic : subset sum