Create The Weakest Wall
By using Binary search concept in this problem, we are traversing through entire length of array for checking no. of subsets & binary search takes over the range [max-1,sum] .
So, the time complexity of this algorithm is roughly O(n.log(sum)) . 

If we use naive approach, it takes O(n*n) for which TLE occurs.